requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
gunicorn==20.1.0
Werkzeug==2.0.1
MarkupSafe==2.0.1
Jinja2==3.0.1
pytz==2023.3
fastapi==0.110.0
uvicorn==0.28.0
Pillow==10.0.0
python-multipart==0.0.6
# Pillow - instalar manualmente si es necesario 